// src/shared/api/models/getApiInsightsWeeklyTurnoverType200DataWeeksItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiInsightsWeeklyTurnoverType200DataWeeksItemMonday } from './getApiInsightsWeeklyTurnoverType200DataWeeksItemMonday';
import type { GetApiInsightsWeeklyTurnoverType200DataWeeksItemTuesday } from './getApiInsightsWeeklyTurnoverType200DataWeeksItemTuesday';
import type { GetApiInsightsWeeklyTurnoverType200DataWeeksItemWednesday } from './getApiInsightsWeeklyTurnoverType200DataWeeksItemWednesday';
import type { GetApiInsightsWeeklyTurnoverType200DataWeeksItemThursday } from './getApiInsightsWeeklyTurnoverType200DataWeeksItemThursday';
import type { GetApiInsightsWeeklyTurnoverType200DataWeeksItemFriday } from './getApiInsightsWeeklyTurnoverType200DataWeeksItemFriday';
import type { GetApiInsightsWeeklyTurnoverType200DataWeeksItemSaturday } from './getApiInsightsWeeklyTurnoverType200DataWeeksItemSaturday';
import type { GetApiInsightsWeeklyTurnoverType200DataWeeksItemSunday } from './getApiInsightsWeeklyTurnoverType200DataWeeksItemSunday';
import type { GetApiInsightsWeeklyTurnoverType200DataWeeksItemWeekTotal } from './getApiInsightsWeeklyTurnoverType200DataWeeksItemWeekTotal';

export type GetApiInsightsWeeklyTurnoverType200DataWeeksItem = {
    weekNumber: number;
    weekStartDate: string;
    weekEndDate: string;
    monday: GetApiInsightsWeeklyTurnoverType200DataWeeksItemMonday;
    tuesday: GetApiInsightsWeeklyTurnoverType200DataWeeksItemTuesday;
    wednesday: GetApiInsightsWeeklyTurnoverType200DataWeeksItemWednesday;
    thursday: GetApiInsightsWeeklyTurnoverType200DataWeeksItemThursday;
    friday: GetApiInsightsWeeklyTurnoverType200DataWeeksItemFriday;
    saturday: GetApiInsightsWeeklyTurnoverType200DataWeeksItemSaturday;
    sunday: GetApiInsightsWeeklyTurnoverType200DataWeeksItemSunday;
    weekTotal: GetApiInsightsWeeklyTurnoverType200DataWeeksItemWeekTotal;
};